Scorelist with ArrayList V

+++ src/main/java/AgChSaJo/IPlayer.java
@@ -5,5 +5,5 @@ package AgChSaJo;
   */
 public interface IPlayer {
     String getNickname();
-    double getFinalScore();
+    int getFinalScore();
 }
